requirements.txt
requests
beautifulsoup4
pycrypto
cryptography
python-dateutil
urllib3
pyyaml
